requirements.txt
sanic
python-socketio
nvidia-ml-py3
psutil